requirements.txt
numpy==1.23.5
psutil==5.9.5
pyworld==0.3.2
faiss-cpu==1.7.3
monotonic_align==1.0.0
fairseq==0.12.2
scipy==1.9.3
onnxruntime-gpu==1.15
onnxruntime==1.15
torchcrepe==0.0.20
ffmpy==0.3.1
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
fastapi
uvicorn
python-dotenv
boto3
pydantic
pydub